requirements.txt
simpletransformers==0.47.2
newsapi-python==0.2.6
giphy-client==1.0.0
tweepy==3.9.0
Pillow==6.2.2
torch==1.5.0
torchcontrib==0.0.2